fix(main/binutils-libs): append `-Wl,--undefined-version` to LDFLAGS (#18813)
Saw this in #18758. llvm appears to have upgraded a warning to an error
(https://reviews.llvm.org/D135402), and while they're not wrong to do
so, this package and others
(https://github.com/rust-lang/rust/issues/105967) depended on it. This
is the simple fix, per the llvm reviews link.

Fixes #18750

[skip ci]